Update dothey.md
Tightened up several wordy things here and rephrased some things that sounded clunky or repetitive.  You also have a habit of including some meandering thoughts at the conclusion of a lot of your paragraphs and sections that are just anecdotes or don't belong, and this weakens the impact of the points you just made.  I would work on avoiding that.

More notes:

Line 3:  I added some more illustrative text to make the idea sound more "magical"

Line 5: Changed "ten" to "several".  Ten is not an impressive enough number to call out.  Also, I'm not entirely sure if what you meant by "they are the loud ones" is "they are the opinionated ones", but that's what I went with.

Last line of main section:  Removed.  This ditracts from your point that there's a perception/reality divide sometimes.

Third and fourth lines of "It depends":  Removed.  You cover this better in the "age" section.  It is unnecessary here and doesn't segue well into the four lifestyle factor points.  

Lines 21-24: Made bullet statements a little clearer with leading categories.

Last line of "age of programmers" section: Removed.  This is highly debatable depending on what kind of grandparents you had (or if they're even alive) and doesn't add anything to your point.

Line 57:  Changed the title of this a little to transition better.

Line 67:   I think it's important to mention that this only gets important if they're earning money with it.

Last two lines of "Lifestyle" section:  Removed.  I have no idea what you're trying to say here.  What folk wisdom?  What statistics?  It's called a day job because it's been historically impractical for humans to conduct business at a time when we can't see what we're doing, so we centered our society around when the sun would be out.

"What about other creatives?" section:  I'm not entirely sure what to do with this section, but I don't think it belongs here because it interrupts your message by essentially saying "I don't want to do the research to talk about you other creative types".  You might include it as a footnote to your introduction or something.

"Statistics" section: the notes indicate that there's more you want to add here, so I haven't touched it yet.